created useCart hook, refactored CatalogBookCard
