+++ frontend/src/pages/CatalogPage/CatalogBookCard/CatalogBookCard.tsx
@@ -1,20 +1,14 @@
-import { images } from "assets/images";
+import { useCallback, useMemo, useState } from "react";
+import { useNavigate } from "react-router-dom";
+import ReactStars from "react-rating-stars-component";
+
 import Cart from "components/Cart";
 import FavoriteButton from "components/FavoriteButton/";
 import Modal from "components/Modal";
-
-import { useCallback, useMemo, useState } from "react";
-import ReactStars from "react-rating-stars-component";
-import { useNavigate } from "react-router-dom";
+import useCart from "hooks/useCart";
 import { truncateString } from "utils/truncateString";
 import { useAppDispatch } from "../../../redux/hooks";
-import {
-  addToCart,
-  createCart,
-  fetchOrderById,
-  fetchOrdersData,
-} from "../../../redux/orders/operations";
-
+import { images } from "assets/images";
 import {
   ButtonOrange,
   ButtonYellow,
@@ -31,7 +25,6 @@ import {
   StyledStarsWrapper,
   StyledTitle,
 } from "./CatalogBookCard.styled";
-
 import { StyledStarIcon } from "components/BookRating/BookRating.styled";
 
 interface IProps {
@@ -52,8 +45,9 @@ const BookCard: React.FC<IProps> = ({
   price,
   rating,
 }) => {
-  const dispatch = useAppDispatch();
+
   let navigate = useNavigate();
+  const { handleCart } = useCart(_id);
 
   const starsProps = useMemo(
     () => ({
@@ -93,72 +87,19 @@ const BookCard: React.FC<IProps> = ({
     if (!localStorage.getItem(`isBookAdded_${_id}`)) {
       localStorage.setItem(`isBookAdded_${_id}`, "true");
 
-      let orderId = localStorage.getItem("currentOrderId");
-
-      if (!orderId) {
-        const createCartResponse = await dispatch(createCart());
-        console.log(createCartResponse);
-        if (createCartResponse.meta.requestStatus === "fulfilled") {
-          orderId = createCartResponse?.payload?.orderId;
-          if (orderId) {
-            localStorage.setItem("currentOrderId", orderId);
-          } else {
-            console.log("orderId не встановлений");
-            return;
-          }
-        } else {
-          console.log("Помилка при створенні кошика.");
-          return;
-        }
-      }
-      const addToCartResponse = await dispatch(
-        addToCart({ orderId, productId: _id })
-      );
-      if (addToCartResponse.meta.requestStatus === "fulfilled") {
-        await dispatch(fetchOrderById(orderId));
-        localStorage.setItem(`isBookAdded_${_id}`, "true");
-      } else {
-        console.log("Помилка при додаванні товару до кошика.");
-      }
+      await handleCart();
     }
 
     showModal("cart");
-  }, [_id, dispatch]);
+  }, [_id, handleCart]);
 
   const handleAddToCart = useCallback(async () => {
     if (localStorage.getItem(`isBookAdded_${_id}`)) {
       showModal("isBookAdded");
     } else {
-      let orderId = localStorage.getItem("currentOrderId");
-
-      if (!orderId) {
-        const createCartResponse = await dispatch(createCart());
-        console.log(createCartResponse);
-        if (createCartResponse.meta.requestStatus === "fulfilled") {
-          orderId = createCartResponse?.payload?.orderId;
-          if (orderId) {
-            localStorage.setItem("currentOrderId", orderId);
-          } else {
-            console.log("orderId не встановлений");
-            return;
-          }
-        } else {
-          console.log("Помилка при створенні кошика.");
-          return;
-        }
-      }
-
-      const addToCartResponse = await dispatch(
-        addToCart({ orderId, productId: _id })
-      );
-      if (addToCartResponse.meta.requestStatus === "fulfilled") {
-        await dispatch(fetchOrderById(orderId));
-        localStorage.setItem(`isBookAdded_${_id}`, "true");
-      } else {
-        console.log("Помилка при додаванні товару до кошика.");
-      }
+      await handleCart();
     }
-  }, [_id, dispatch]);
+  }, [_id, handleCart]);
 
   return (
     <>
